abstract validateTargetAccess.ts to reuse
